refactor(instrumentation-redis): support multi versions as seprate inst objects
