+++ plugins/node/opentelemetry-instrumentation-redis/src/v1-2-3/internal-types.ts
